backend/requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.0
openai>=1.12.0
langchain>=0.3.0,<0.4.0
langchain-core>=0.3.0
langchain-community>=0.3.9
langchain-openai>=0.1.0
langchain-pinecone>=0.1.0
langchain-text-splitters>=0.1.0
pinecone>=3.0.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
pypdf>=5.3.0
pandas>=2.0.0
python-multipart>=0.0.6